+++ strider/merge_tracks/__init__.py
@@ -1,4 +1,5 @@
-from strider.merge_tracks.rules import rule, action, trigger,\
-    source_re, add_tags, change_id, has_tag,\
-    SKIP
-from strider.merge_tracks.load_rules import load_rule_file
+from strider.merge_tracks.rules import (rule, action, trigger,
+                                        source_re, add_tags, change_id, has_tag, default,
+                                        SKIP,
+                                        Rule, Trigger, Action)
+from strider.merge_tracks.load_rules import load_rule_file, load_inline_rule